requirements.txt
apache-airflow[amazon]